Make sort optional in XChildren() method.
In some cases we don't need the child nodes returned by the XChildren() API to be sorted, and as sorting is an expensive operation, it is useful to be able to avoid this overhead. This commit adds a new parameter to XChildren() - sortSpec - that allows the caller to request whether the returned nodes are sorted in XPath order or not.

I've also moved one make() call to avoid it being called unnecessarily.

Part of VRVDR-53594. Needs to be committed at same time as configd and golang-github-danos-xpath-plugins changes.

(cherry picked from commit f5ec0f5b19354ef225260db29e10a79f03f56db7)